Fixed JS
I’m not sure why they coded it like that, but something was interfering
with the function from running. I changed it to code that I’m more
familiar with. Now it is correctly adding the “in-view” class. The
animation still isn’t working for me. See if you can find out why and
I’ll try to find it as well.