Added subject class
In the style.css You can see it is trying to add the animation to
something with the class: “subject”. The period in front means it is a
class. You can see in style.css on line 796. It needed to be within
your div with the class animation-element. So I knew I just needed to
add the class subject to the img tag and it would work.